Gulp distribution build fixes
Removed non-local script sources out of build block. Fixed font urls in
css.

Added ignore regex to Uncss pipe, so that .spinner classes for the
loading directive would not be removed. This is because the 'loading'
directive in index.html doesn't reference .spinner directly, so Uncss
thinks it isn't used.